requirements.txt
streamlit
faiss-cpu
sentence-transformers
huggingface_hub
pandas
numpy
requests
streamlit_lottie
